Move EnhancedButton under internal
