+++ src/floating-action-button.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Transitions from './styles/transitions';
 import ColorManipulator from './utils/color-manipulator';
-import EnhancedButton from './enhanced-button';
+import EnhancedButton from './internal/EnhancedButton';
 import FontIcon from './font-icon';
 import Paper from './paper';
 import Children from './utils/children';
