+++ src/icon-button.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Transitions from './styles/transitions';
 import PropTypes from './utils/prop-types';
-import EnhancedButton from './enhanced-button';
+import EnhancedButton from './internal/EnhancedButton';
 import FontIcon from './font-icon';
 import Tooltip from './tooltip';
 import Children from './utils/children';
